Third commit
Remove extraneous test functions, debug logs.
Major changes:
Added logic to implement info box, mapper
Hooked the mobile food link  to a mapper on the map
constructed a new API call to look up the food truck with an unique id
and return the co-ordinates and map it on the map.